Split migrators into files.
